requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
discord.py==2.2.3
fonttools==4.39.4
frozenlist==1.3.3
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
lxml==4.9.2
mariadb==1.0.11
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
PublicDataReader==1.0.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
xmltodict==0.13.0
yarl==1.9.2
zipp==3.15.0
